+++ Examples/RudderSampleAppObjC/RudderSampleAppObjC/_AppDelegate.m
@@ -8,43 +8,38 @@
 
 #import "_AppDelegate.h"
 #import <Rudder/Rudder.h>
+#import "RudderAmplitudeFactory.h"
+#import "RudderBrazeFactory.h"
 #import <AdSupport/ASIdentifierManager.h>
-#import "CustomFactory.h"
 
 
-static NSString *WRITE_KEY = @"21zVhiRJL38EAgphqL65VpzyjLB";
+static NSString *WRITE_KEY = @"2CpkPQoOX97OkRqVemj5PvUdvkg";
+static NSString *DATA_PLANE_URL = @"https://shadowfax-dataplane.dev-rudder.rudderlabs.com";
+static NSString *CONTROL_PLANE_URL = @"https://api.dev.rudderlabs.com";
 
 @implementation _AppDelegate
 
-NSString *const kGCMMessageIDKey = @"gcm.message_id";
-
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions
 {
-    
-    [RSClient putDeviceToken:@"your_device_token"];
-    [RSClient putAnonymousId:@"anonymous_id"];
-    
     
     RSConfigBuilder *builder = [[RSConfigBuilder alloc] init];
     [builder withLoglevel:RSLogLevelVerbose];
     [builder withTrackLifecycleEvens:YES];
     [builder withRecordScreenViews:YES];
-    [builder withEnableBackgroundMode:YES];
-    [builder withDataPlaneUrl:@"https://7b7a-61-95-158-116.ngrok.io"];
-    [RSClient getInstance:@"21zVhiRJL38EAgphqL65VpzyjLB" config:[builder build]];
+    [builder withDataPlaneUrl:DATA_PLANE_URL];
+    [builder withControlPlaneUrl:CONTROL_PLANE_URL];
+    [builder withFactory:[RudderAmplitudeFactory instance]];
+    [builder withFactory:[RudderBrazeFactory instance]];
+    [RSClient getInstance:WRITE_KEY config:[builder build]];
     
     [[RSClient sharedInstance] track:@"simple_track_with_props" properties:@{
         @"key_1" : @"value_1",
         @"key_2" : @"value_2"
     }];
     
-    [[[RSClient sharedInstance] getContext] putAdvertisementId:@"advertisement_Id"];
-    
-    RSOption *identifyOptions = [[RSOption alloc] init];
-    [identifyOptions putExternalId:@"brazeExternalId" withId:@"some_external_id_1"];
     [[RSClient sharedInstance] identify:@"testUserId"
                                  traits:@{@"firstname": @"First Name"}
-                                options:identifyOptions];
+                                options:nil];
     
     [[RSClient sharedInstance] screen:@"ViewController"];
     
@@ -55,28 +50,25 @@ NSString *const kGCMMessageIDKey = @"gcm.message_id";
     ];
     
     [[RSClient sharedInstance] alias:@"new_user_id"];
-    [FIRApp configure];
-    [FIRMessaging messaging].delegate = self;
     
-    [UNUserNotificationCenter currentNotificationCenter].delegate = self;
-    UNAuthorizationOptions authOptions = UNAuthorizationOptionAlert |
-    UNAuthorizationOptionSound | UNAuthorizationOptionBadge;
-    [[UNUserNotificationCenter currentNotificationCenter]
-     requestAuthorizationWithOptions:authOptions
-     completionHandler:^(BOOL granted, NSError * _Nullable error) {
-        // ...
-    }];
+    RSOption* option1 = [[RSOption alloc] init];
+    [option1 putIntegration:@"Amplitude" isEnabled:YES];
     
-    [application registerForRemoteNotifications];
+    RSOption* option2 = [[RSOption alloc] init];
+    [option2 putIntegration:@"Braze" isEnabled:YES];
+    
+    NSDictionary* props = @{@"data": @YES};
+    
+    for(int i=0; i<5;i++) {
+        NSDictionary* duplicateProps = nil;
+        if(i%2 ==0){
+            duplicateProps = props;
+        }
+        [[RSClient sharedInstance] track:[[NSString alloc] initWithFormat:@"Test Event %@: %d", @"Amplitude", i] properties:props options:option1];
+        [[RSClient sharedInstance] track:[[NSString alloc] initWithFormat:@"Test Event %@: %d", @"Braze", i] properties:props options:option2];
+    }
+    
     return YES;
-}
-
-- (NSString*) getIDFA {
-    return [[[ASIdentifierManager sharedManager] advertisingIdentifier] UUIDString];
-}
-
-- (NSString*) getDeviceToken {
-    return @"example_device_token";
 }
 
 - (void)applicationWillResignActive:(UIApplication *)application
@@ -105,100 +97,4 @@ NSString *const kGCMMessageIDKey = @"gcm.message_id";
 {
     // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
 }
-
-// [START receive_message]
-- (void)application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo
-fetchCompletionHandler:(void (^)(UIBackgroundFetchResult))completionHandler {
-    // If you are receiving a notification message while your app is in the background,
-    // this callback will not be fired till the user taps on the notification launching the application.
-    // TODO: Handle data of notification
-    
-    // With swizzling disabled you must let Messaging know about the message, for Analytics
-    // [[FIRMessaging messaging] appDidReceiveMessage:userInfo];
-    
-    // [START_EXCLUDE]
-    // Print message ID.
-    if (userInfo[kGCMMessageIDKey]) {
-        NSLog(@"Message ID: %@", userInfo[kGCMMessageIDKey]);
-    }
-    // [END_EXCLUDE]
-    
-    // Print full message.
-    NSLog(@"%@", userInfo);
-    
-    completionHandler(UIBackgroundFetchResultNewData);
-}
-// [END receive_message]
-
-// [START ios_10_message_handling]
-// Receive displayed notifications for iOS 10 devices.
-// Handle incoming notification messages while app is in the foreground.
-- (void)userNotificationCenter:(UNUserNotificationCenter *)center
-       willPresentNotification:(UNNotification *)notification
-         withCompletionHandler:(void (^)(UNNotificationPresentationOptions))completionHandler {
-    NSDictionary *userInfo = notification.request.content.userInfo;
-    
-    // With swizzling disabled you must let Messaging know about the message, for Analytics
-    // [[FIRMessaging messaging] appDidReceiveMessage:userInfo];
-    
-    // [START_EXCLUDE]
-    // Print message ID.
-    if (userInfo[kGCMMessageIDKey]) {
-        NSLog(@"Message ID: %@", userInfo[kGCMMessageIDKey]);
-    }
-    // [END_EXCLUDE]
-    
-    // Print full message.
-    NSLog(@"%@", userInfo);
-    
-    // Change this to your preferred presentation option
-    completionHandler(UNNotificationPresentationOptionBadge | UNNotificationPresentationOptionAlert);
-}
-
-// Handle notification messages after display notification is tapped by the user.
-- (void)userNotificationCenter:(UNUserNotificationCenter *)center
-didReceiveNotificationResponse:(UNNotificationResponse *)response
-         withCompletionHandler:(void(^)(void))completionHandler {
-    NSDictionary *userInfo = response.notification.request.content.userInfo;
-    if (userInfo[kGCMMessageIDKey]) {
-        NSLog(@"Message ID: %@", userInfo[kGCMMessageIDKey]);
-    }
-    
-    // With swizzling disabled you must let Messaging know about the message, for Analytics
-    // [[FIRMessaging messaging] appDidReceiveMessage:userInfo];
-    
-    // Print full message.
-    NSLog(@"%@", userInfo);
-    
-    completionHandler();
-}
-
-// [END ios_10_message_handling]
-
-// [START refresh_token]
-- (void)messaging:(FIRMessaging *)messaging didReceiveRegistrationToken:(NSString *)fcmToken {
-    NSLog(@"FCM registration token: %@", fcmToken);
-    // Notify about received token.
-    NSDictionary *dataDict = [NSDictionary dictionaryWithObject:fcmToken forKey:@"token"];
-    [[NSNotificationCenter defaultCenter] postNotificationName:
-     @"FCMToken" object:nil userInfo:dataDict];
-    // TODO: If necessary send token to application server.
-    // Note: This callback is fired at each app startup and whenever a new token is generated.
-}
-// [END refresh_token]
-
-- (void)application:(UIApplication *)application didFailToRegisterForRemoteNotificationsWithError:(NSError *)error {
-    NSLog(@"Unable to register for remote notifications: %@", error);
-}
-
-// This function is added here only for debugging purposes, and can be removed if swizzling is enabled.
-// If swizzling is disabled then this function must be implemented so that the APNs device token can be paired to
-// the FCM registration token.
-- (void)application:(UIApplication *)application didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken {
-    NSLog(@"APNs device token retrieved: %@", deviceToken);
-    
-    // With swizzling disabled you must set the APNs device token here.
-    // [FIRMessaging messaging].APNSToken = deviceToken;
-}
-
 @end