+++ Sources/Classes/Public/RSDBPersistentManager.h
@@ -33,20 +33,16 @@ NS_ASSUME_NONNULL_BEGIN
 -(void) createDB;
 -(void) createTables;
 -(void) createEventsTableWithVersion:(int) version;
--(void) createEventsToDestinationIdMappingTable;
 -(void) checkForMigrations;
 -(BOOL) checkIfStatusColumnExists;
 -(void) performMigration;
 -(NSNumber*) saveEvent: (NSString*) message;
+- (void) clearOldEventsWithThreshold:(int)threshold;
 -(void) clearEventsFromDB: (NSMutableArray*) messageIds;
 -(RSDBMessage *)fetchEventsFromDB:(int)count ForMode:(MODES) mode;
 -(RSDBMessage*) fetchAllEventsFromDBForMode:(MODES) mode;
 -(void) updateEventsWithIds:(NSMutableArray*) messageIds withStatus:(EVENTPROCESSINGSTATUS) status;
 -(void) clearProcessedEventsFromDB;
--(NSArray<NSString*>*) getEventIdsWithDestinationMapping:(NSArray*) eventIds;
-- (void) saveEvent:(NSNumber*) rowId toDestinationId:(NSString*) destinationId;
-- (void) deleteEvents:(NSArray*) eventIds withDestinationId:(NSString*) destinationId;
--(NSDictionary<NSString*, NSArray<NSString*>*>*) getDestinationMappingofEvents:(NSArray<NSString*>*) eventIds;
 - (int) getDBRecordCountForMode:(MODES) mode;
 -(void) flushEventsFromDB;