+++ Sources/Classes/Public/RSUtils.h
@@ -27,6 +27,8 @@ NS_ASSUME_NONNULL_BEGIN
 + (NSString*) getCSVString:(NSArray*) inputStrings;
 + (NSString*) getJSONCSVString:(NSArray*) inputStrings;
 + (id) deSerializeJSONString:(NSString*) jsonString;
++ (BOOL) isValidURL:(NSURL*) url;
++ (NSString*) appendSlashToUrl:(NSString*) url;
 
 extern unsigned int MAX_EVENT_SIZE;
 extern unsigned int MAX_BATCH_SIZE;