+++ Sources/Classes/RSUtils.m
@@ -28,7 +28,6 @@
 + (const char *)getDBPath {
     NSURL *urlDirectory = [[NSFileManager defaultManager] URLsForDirectory:NSLibraryDirectory inDomains:NSUserDomainMask][0];
     NSURL *fileUrl = [urlDirectory URLByAppendingPathComponent:@"rl_persistence.sqlite"];
-    NSLog(@"Path of the db desu is %@", [fileUrl path]);
     return [[fileUrl path] UTF8String];
 }
 
@@ -168,6 +167,16 @@
         return nil;
     }
     return object;
+
++ (BOOL) isValidURL:(NSURL*) url {
+    return url && [url scheme] && [url host];
+}
+
++ (NSString*) appendSlashToUrl:(NSString*) url {
+    if([url hasSuffix:@"/"]){
+        return url;
+    }
+    return [url stringByAppendingString:@"/"];
 }
 
 unsigned int MAX_EVENT_SIZE = 32 * 1024; // 32 KB